migrate from game-session into game-sessions
